requirements-cpu.txt
# Streamlit Cloud で通常の requirements.txt が失敗する場合の代替案
# Deploy → Advanced settings → Requirements file でこのファイルを指定
# Python は 3.11 / 3.12 のみ（3.14 不可。理由は requirements.txt のコメント参照）
# pyannote.audio / whisperx を NumPy 2.x と併用しない（np.NaN 削除）
numpy>=1.26,<2
matplotlib>=3.7
faster-whisper==1.0.3
streamlit>=1.36.0
python-dotenv>=1.0.0
openai>=1.0.0
yt-dlp>=2024.1.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
torchvision==0.19.1+cpu

whisperx==3.2.0
